requirements.txt
streamlit
pandas
biopython
